+++ src/pages/Authentication/form/AuthenticationSchemas.tsx
@@ -1,5 +1,5 @@
 import * as Yup from 'yup'
-import { userSchema } from '../User/form/user.schemas'
+import { userSchema } from '../../User/form/user.schemas'
 
 export const signInSchema = Yup.object().shape({
     username: userSchema.username,
